Fixed upgrading of equipped items
It shouldn't be possible to upgrade an equipped item.
The original server allowed this, however people managed to downgrade
their maxed out weapons to +6 when some visual bugs on the client
occured.
Example: On the server side there is a jewel of bless on a certain slot,
but client shows a health potion. When the client then consumes the
potion it would apply the bless to item slot 0.