Changed loading of GameConfiguration objects to use queries which return them as json for better performance
See also #10.
Also prepared to use it for accounts - we still have some minor issues
there. For example an ItemStorage loads it's items, and items load the
item storage which tries to load it's items again... and so on. So the
JsonObjectLoader needs some adjustments to handle such cases.
Another limitation of the new approach for the GameConfiguration is,
that the DbContext can't track the changes anymore. It contains circular
references, so we load this json twice, whereas the first run is just
there to fill the cache of the IdReferenceResolver - leading to
different (but same) instances with the same Id - the change tracker is
not afraid about that ;)