Changed the naming convention of internal join entity collection properties
Instead of using the type name, we use the name of the property. This
way we can conclude from the join entity collection to the target type
property name. For example, "JoinedQualifiedCharacterClasses" ->
"QualifiedCharacterClasses". We are also less likely to get naming
conflicts when we have multiple many-to-many collection properties
between the same types.
Required for a solution of #21.