Replaced GameMapDefinition.DeathSafezone by SafezoneMap
Found a flaw in the data model - GameMapDefinition.SpawnGates included
all ExitGates of a map, but was only intended to include points where
players spawn after joining the map by warping, after dying etc.
So to fix this, I renamed it to ExitGates and added a flag 'IsSpawnGate'
to the ExitGate. I renamed 'Gates' to 'EnterGates' accordingly.