Fixed an issue when loading a previously created character
When a character gets created, it creates the StatAttributes by
indirectly calling the constructor. When the definition field got set,
it didn't come along the overwritten property of the ef core class and
didn't set the DefinitionId, so that after loading a character it had
StatAttributes without AttributeDefinitions.
Setting the property instead of the field fixes this problem.