A major refactoring of the guild server and the data handling
Oh, this commit got a little bit too big... It all started with making
the ChatServer independent from the DataModel. I moved some model
classes to the Interfaces project and noticed that the guild server
interface and its data model was a bit messy and inconsistent.

The guild server is now more independent - the interface cleaner so that
it doesn't need to send data model classes around.
Then it's more consistent because guilds are referenced by 32bit
integers, characters mainly by their GUID.