A major refactoring of the persistence interface
Repositories are now an implementation detail. The application uses now
"contexts" provided by a PersistenceContextProvider. This makes data
access easier for the application and is easier to implement in the
persistence layer.
Added/Adapted Readmes accordingly.
Tests use now the in memory contexts which makes them easier to set up
as well.
Also replaced the workaround of the "FriendView" by using linq queries
instead which join character names to Friend objects.