Fixed construction of the json query for GameConfiguration
Fixes #38. We move the maps with their spawn points to the end because they depend on all other data. This fixes the problem that the Skill property of merchant items are not set. Actually, the assigned MonsterDefinition instances of the spawn areas were not fully loaded, in this particular case.
I hope this doesn't have other side-effects, though.