Remove Durability from create Learnable
Durability on learnable item causes item remain in the inventory when being used. It will disappear on relog.
But in the same time it is removed from Context so it become acting as bugged item.
Like simply you can grab it in inventory but it is not moving e.t.c.