Fixed bug where player wasn't removed from observers of NPCs
This is related to the dotnetcore branch because I did a change before with https://github.com/MUnique/OpenMU/commit/48644724d26a4f991ea12a32bf1a0cdc39b8166f. However, I'm not sure if it's related to the migration at all.

Anyway... this should fix the problem that the player stays as observer of all met NPCs until it left the map. The old buckets need to be removed before calling LocateablesOutOfScope, because it compares the current bucket (NewBucket) of the item which got out of scope with the currently ObservingBuckets - if we would keep the old buckets, the object would never get removed.
I refactored the code a bit (ObserverToWorldViewAdapter.ObjectWillBeOutOfScope) to be able to better understand it.