Added functionality to add new connection servers at runtime
It's not perfect yet, but a good start.
To make it useful, we also need the following stuff:
  * Possibility to add more game client definitions
  * Possibility to add game server endpoints for specific game clients
  * Possibility to add game servers for specific game configurations

Also need to think about rethinking all of the server state listeners - there are now more than one ;)

In the big picture these feature also contribute to the plugin system, because plugins are active for specific game client versions. See also #61.