Introduced Minimum- and MaximumClientAttribute
This is required to prevent some error messages and misbehaviors when connecting with a new client when there are same kind of plugins with same or different keys for lower versions. It may otherwise happen that these lower level plugins are initialized before the higher version ones. In this case there would be a error message in the log.
Now, we can define the minimum and maximum client version for each plugin type. When these versions are known, it's recommended to add Attributes for them.