Added basic support for version 0.75 (#114)
* Added client definition for version 0.75

* Fixed chaos mix packet handler

  * index 2 always contains 0x86, so removed this check
  * the id of the mix is at index 3, and if it's not available (in case of version 0.75), we just use id 1 (chaos weapon mix)

* ItemMoveHandlerPlugIn independent from client version

The item size is now determined by asking the item serializer which depends on the client version.

* Removed redundant assignment of the packet length

* corrected comment

The two missing bytes are padding, because the "current health" is an integer, so it can only start every 4 bytes (C-style struct alignment).

* Added some guild join results and missing checks

* The guild kick success message needs to be sent to the affected player

* Consolidated two similar plugin interfaces to one

One view plugin should only implement one view plugin interface

* Fixed call to consolidated plugin interface

* clarified the meaning of the 3rd byte - padding

* removed unused using

* Fixed wrong interface in CharacterCreatePacketHandlerPlugIn

It needs to implement ISubPacketHandlerPlugIn to get picked up by the PacketGroupHandlerPlugIn

* Made character deletion more compatible with other versions

* Added vault extension flag to character list packet

* Added check in character deletion

Guild members are not allowed to get deleted - required check for 0.75. They need to be removed from their guild first. Also adapted the character list plugin to show the current guild state - one TODO less :)

* Refactored simple modulus key generator and serializer to support 16 keys

16 keys are used by the version 0.75 client.
See also #109.

* Adapted SimpleModulus to support bigger block sizes

0.75 uses blocks of 32 bytes and no counter

* Simplified reading integers from streams

* added missing comment

* Fixed Exception when encrypting short packets with SimpleModulus

sourceOffset is 0, so minus 1 would lead to an ArgumentOutOfRangeException.
Minus 1 is required when a counter is involved.

* Fixed broken SimpleModulusEncryption

An OutOfRangeException occurs in the bottom loop when we add 1 to totalDecryptedSize.

* The connect server should only handle the span with the actual packet size

* Fixed reading of the server id for older versions

the field size is determined by the packet size

* Added packet handler for the old op-code of the server list request

Instead of 0x06, it uses 0x02. It doesn't hurt to handle both.

* Added client version parameter for the connect server

One connect server instance can handle one specific client protocol.

* Added login handler plugin for 0.75

* Fixed ClientVersion comparison

When both languages are invariant, the should be handled the same.

* Added appearance serializer for version 0.75

* Extracted CharacterAppearanceDataAdapter from parent class

* Added character list plugin for 0.75

* Added plugin for character stats in version 0.75

* Added item serializer and adapted inventory list plugin for version 0.75

not sure yet if it's 100% correct. Some items seem to be missing after login :-(

* Made login window plugin available for version 0.75

It works there without changes.

* Reverted unwanted change

* Added NewPlayersInScopePlugIn for version 0.75

* Flags enum should be named plural

* Added a default switch case with comment

* Implemented packet handler for walking in version 0.75

* Added view plugin for 'NPCs in scope' for version 0.75

* Fixed item pick up in 0.75

It's not encrypted as C3 packet, but C1.

* Made ShowHitPlugIn compatible with various client versions

Including 0.75

* Removed message prefix for early versions

Otherwise, it shows it in the client.

* Fixed inventory list plugin and item serializer

In 0.75 each item is 3 bytes long. It doesn't support excellent options yet.

* Fixed hit packet handler to handle the correct key in 0.75.

* Extended MessageType by GuildNotice

* Merged switch cases

Korean and Filipino clients use the same operation code 0xDF.

* Added character move packet handler for 0.75

* Introduced Minimum- and MaximumClientAttribute

This is required to prevent some error messages and misbehaviors when connecting with a new client when there are same kind of plugins with same or different keys for lower versions. It may otherwise happen that these lower level plugins are initialized before the higher version ones. In this case there would be a error message in the log.
Now, we can define the minimum and maximum client version for each plugin type. When these versions are known, it's recommended to add Attributes for them.

* Fixed consideration of inherited Min/Max-ClientAttributes

The inherited attributes shouldn't be considered. Added tests for that.
Otherwise, some plugins which inherit from other plugins are not considered fitting for any client version anymore.

* Implemented Skills for version 0.75

Adapted skill list message and skill packet handler

* Fixed Guild tests

There were added checks for the player levels, so the test player did never join the guild before this fix.

* Fixed ItemConsumptionTest

There were added checks to be able to only update the correct items - wearable ones. The test has to be adapted, as well.