Implemented an initial approach to generate packet structs by data
For the data, I basically took the XML files of the network analyzer and extended them for the new requirements.
Then I wrote a XSL transformation which transforms this xml data to C# code.
I set up the transformation to run as pre-build event, so that it first runs the transformation, then compiles the code. This way, it gave me a fast feedback-loop to adapt/modify the transformation until I was satisfied with the result.
In this process I also began to go through the ClientToServer packets and extended some missing data - this work is still in progress, a lot of messages are still missing.

The plan is, to use this project in the GameServer project and the Network.Analyzer. Later maybe also in the ChatServer and ExDbConnector.