Fixed T4 template execution
The problem was, that the assemblies of OpenMU were directly referenced in the T4 templates to analyze their types etc.
This is a problem, since we moved them all to .net standard 2.1 and the runtime which executes these templates is probably the normal .net framework (used by Visual Studio) which doesn't implement the standard 2.1.
So, to fix this, I changed the templates to load the assemblies to use Assembly.ReflectionOnlyLoadFrom instead. This "somehow" works, but I had to implement an event handler to resolve the referenced assemblies, especially the netstandard.dll.

I'm happy, that it works now, because this was the only way I found - and I put a lot of effort into it.