Disabled the XSL transformation in a CI environment
The transformation works when working with visual studio 2017, but unfortunately not in cases where only the .net core sdk is available. See also https://github.com/Microsoft/msbuild/issues/3368.

So, to keep CI build running, the transformation was disabled. The generated source code should be checked in with the xml files anyway.