Added packet definition for the NewPlayersInScopePlugIn
I had to extend the data model of the packet definition, so that it's possible to extend structs by a custom implementation of the indexer.
The size of each character block may differ and is not constant through a packet. So, the code has to go through the packet to see where the n-th block is