Improved the initialization of the MapDefinition.SafezoneMap
It sets now the map itself as SafezoneMap, if the map has a safezone.
If not, Lorencia is set as a default.
This should work in the most cases, but of course there are some maps which don't take you back to Lorencia, e.g. some events take you back to Devias or Noria. These will need to be adapted in the future. For now, it's okay to just use Lorencia.
See #146.