Fixed InvalidOperationException when a dropped item is deleted
We don't need to track all navigation properties of an item. The item itself is enough - that's why for now, the TradeContext is sufficient.

The exception was: System.InvalidOperationException: The instance of entity type 'ExitGate' cannot be tracked because another instance with the same key value for {'Id'} is already being tracked. When attaching existing entities, ensure that only one entity instance with a given key value is attached. Consider using 'DbContextOptionsBuilder.EnableSensitiveDataLogging' to see the conflicting key values.