Implemented the game live map (with issues)
Unfortunately, objects are not yet shown on the map. I couldn't figure out the root of the cause yet. From my observations, it should work; The calls to add the objects are arriving in the maps javascript code. The code is almost the same as in the old AdminPanel, so I'm a bit clueless. Increasing the size of the 3d objects and making them intransparent didn't make them visible, too.