Fixed picking up items in Noria
The game client just supports a certain amount of dropped items. I think it uses an array internally. On the original server, ID ranges of items and other objects are disjoint. When the drop id is higher than the array is long, it wont be able to pick the items up again. To solve this, we use the first 0x1FF IDs in a map exclusively for items. This allows us to keep the simple implementation of one dictionary for all objects in a map.

See #146.