Implemented a very simple generic page to edit data
alpha-status - contains bugs!
It shows properties as input fields, based on the reflected properties of the given type. As it makes heavy use of reflection, the output might not fit to every case to 100%.

I decided to use Blazored.Typeahead to show object lookups and list of objects. The control might not always be the right one to use for all kind of lists. It also has an issue when clicking on the link and it just binds to lists, not collections. I'm finding a way to sort this out soon.

When the save button was clicked, there is no feedback about the success yet.