Improvements for the AutoFields and ItemTable components
The add or create button is shown, depending if the property is a member of the aggregate or not.
The ItemTable is now shown as bootstrap card instead of a jumbotron - this looks much better and underlines the structure of the data.
For properties which are a member of the aggregate, but not a collection, I introduced the MemberOfAggregateField. It inlines all fields of it, if it's set. If it's null, it shows a button to create the object.