Added extension methods for IConnection to simplify packet sending
They're auto-generated and based on the xml data.
Had to slightly adapt the Type pattern for enums, so that the class is specified in their type name. That's required because these names are not used in the same class, but in extensions.
I also had to reference MUnique.OpenMU.Network, because of the IConnection interface. They were both unrelated, but if I wanted to make use of extension methods I had the following options:
  * Put the extension methods into Network and reference Network.Packets
  * Put the extension methods into Network.Packets and reference Network (-> I did this)
  * Create a new project which references both

For the sake of simplicity I've decided to run all xslt in Network.Packets.

Bumped Network.Packets nuget version to 0.3.4