Added edit pages for servers
In this process I realized that I had a big flaw in my persistence layer. I refactored it, so that the admin panel shows freshly loaded data from the database, instead of directly editing cached data which is used by the live server. This also led to some other tricky problems, because objects were in different contexts.
Now, what's probably missing is the refresh of the loaded configuration at the servers. However, now we don't have unwanted side-effects anymore, so this was worth the effort.