Fixed setting object references through the lookup control
It needs to use the same persistence context, otherwise we get problems when we try to save the object which was modified.
Example, when changing the CurrentMap of a Character: "An unexpected error occured: The instance of entity type 'GameMapDefinition' cannot be tracked because another instance with the same key value for {'Id'} is already being tracked. When attaching existing entities, ensure that only one entity instance with a given key value is attached. Consider using 'DbContextOptionsBuilder.EnableSensitiveDataLogging' to see the conflicting key values.."

Thanks @fos83 for the bug report