Fixed login issue
When you entered a false password the first time, it wouldn't let you in even if you entered the correct one in a second attempt.
The reason is, that an exception happend because the account object was attached to the DbContext in the first try - attaching it a second time with a different object instance throws an exception.
The fix:
  * removed one attachment, so it happens now only in GetId
  * added a different login result, so that's clear that it's not the username/password which is wrong in case an exception happend
  * minor code cleanups