Fixed starting of quests
Introduced a "starting number" which is basically the number which is reported as available quest and used by the client to start the quest. The regular quest number is used after that state. When selecting a quest, the player is asked about if he wants to start it - previously it went straight to the requirements.

Thanks @fos83 for reporting the issue.