Fixed the quest dialog message when refusing a quest
Instead of going back to the first screen of the dialog, the game client should show a special message about refusing the quest. Another number than the StartingNumber or the quest Number must be sent - this was a hard nut to crack, because I didn't find these numbers anywhere on other implementations. It had to be the number which is missing between the last quest Number of the same StartingNumber and the next StartingNumber - in most of the cases.
I also renamed the StartQuest message to SelectQuest, because it was never starting quest.