Fixed guild issues
When a guild gets deleted, its short id shouldn't be reused right again. Game clients may still cache some guild information for this id. So, the best solution is, to postpone the reusage of the id to the last possible moment.

Having no numbers available is more like a theoretical issue, because a server first has to gain such an amount of guilds (65k+). However, it's not impossible - on GMO they had this issue, but they also used an id for every existing guild of a realm, also famously known as "Guild Bug" like 10 years ago.
OpenMU follows another approach: it just assigns these ids to guilds during the time when their members are actually online.