Fixed possible account loading issue
An MemberOfAggregateAttribute was missing at Character.QuestStates. This had the effect that the collection had a null element in it after loading it though the json query.
When attaching the object graph to the context, it was throwing an exception, because the key (id) was given, but no corresponding object as value.