Fixed InvalidOperationException because of modified collection
When the players are iterated and players are leaving/joining the game, it came to an InvalidOperationException.
This can be prevented by iterating with a simple for-loop, preferrably backwards.
With this fix, it's not guaranteed that every player gets the message or gets it only once. However, this is not really important for this plugin.