catch exceptions when sending network packets
Logging the exceptions is all we can do.
They should not bubble up to the game logic, because the logic might send packets to many clients in a loop - one erroneous connection should not lead to effects on others.
Alternatively, we would have to catch them in the view plugins - with a lot of repeating boilerplate code.