Simplified API of ChatServer and FriendServer
The idea is, that the auth data is just going through one interface. We don't want it as return value, so we can switch to an asynchronous communication model.
Also, removed the IP address from the IChatServer and IFriendServer interfaces, and added it to the ChatServerAuthenticationInfo. This way, we don't have to query these interfaces again for information we'll need by 100 %. This also increases the chances that we can possibly have multiple chat servers in the future - added a todo in FriendServer about that.