Fixed an issue in the model generator code
When an inherited base class has a collection with a standalone type which would cause the generation of a join-class for a many-to-many-relationship, it didn't create the required join class. It created the join class for the inherited class twice. Instead of taking names from the declared type (the base class), we must use the reflected type (the actual, inheriting class).