Box of luck mechanism (See #231) (#236)
* Fixed an issue in the model generator code

When an inherited base class has a collection with a standalone type which would cause the generation of a join-class for a many-to-many-relationship, it didn't create the required join class. It created the join class for the inherited class twice. Instead of taking names from the declared type (the base class), we must use the reflected type (the actual, inheriting class).

* Extension of the data model

* Extension of the drop logic

* Allow more than one ItemDropItemGroup per item and money drops

* Extended packet definitions by some item drop effects

* Added view plugins to show effects

* Added chat commands to test effects

* Added data initialization for box of luck and similar

* Adapted logic to consider effects

* fixed data initialization

* Fixed duplicate drops at 'Blue Ribbon Box'
* Sorted items to spot duplicates easier

* removed some empty lines

* Added Rings of Warrior to new characters

* Removed unused parameter