Fix for item crafting
Sort the required items by the amount, so that optional items are checked last.
Example where this failed previously: When making first wings, it requires a chaos weapon. Optionally, random items can increase the success rate. In some cases, the chaos weapon was recognized as a random item, and not as the required chaos weapon. To fix this, we try to find the non-optional items first.