Fixed admin panel menu on first load
When the admin panel is hosted standalone (daprized), the first game configuration loading will take some time, because it's not in the cache.
The loading logic was a bit flawed, because it didn't signal a state change when loading was done. Additionally, loading was triggered more than once - however, the second load just returned null (see change in CachedRepository), because the first load call wasn't finished yet.