The ClientLauncher supports domain names now
It seems like the game client can only work with IPs written into the registry.
For convenience, the launcher can work with domain names now, too. It will resolve the domain name to an ipv4 address before launching the game client.
It will also handle the case of an entered "127.0.0.1" which the game client blocks - it's replaced behind the scenes by 127.127.127.127.