Fixed Force Wave skill
It hits now multiple targets and is sent to the client in a proper way - only as 'Force' skill. From the client, we also only get the 'Force' (nr, 60) skill when performing it, so we have to check, if a weapon with the force wave skill is equipped and act accordingly.
Since this strange behavior is done by the client, I put corresponding correction logic into the view and packet handler plugins.
Additionally, i noticed that the frustum calculation was always done from the players coordinate - however, this led to unwanted hits. Instead, the hit box starts now right in front of the character.