Optimizations regarding path finding
First, the path finding itself was optimized to only search in a path in a scoped area of a map. For the most searches, a 8x8 field is enough. In the cases where it's not, the field is extended to 16x16. Before this optimization, the whole map of 256x256 was used, which took considerable memory and initialization time.
Next, I implemented pooling for the pathfinder objects. Before, the monster instances each had one instance of a PathFinder object. However, they use it pretty infrequently - even when players are around, just every few seconds. This had a very negative effect on the memory usage. Since a search for a path doesn't need a lot of time - in release build and with optimization above, averaging at just 0,025 ms! - very few PathFinder objects can serve thousands of monsters.