Fixed deleting of items
When we delete new items twice, something strange happens:
After the first delete, the entity entry has the state Detached. So far, so good.
After the second delete, the entity entry has the state Deleted. That's a problem, because then the ef core tries to delete it from the database, where it never existed.