Extensions of the Attribute System
Made it possible to use another attribute as operand, instead of a constant float value. This will make some calculations easier, e.g. we can use it for conditional bonuses. Example: Give a bonus only if an item with a specific attribute is equipped. This comes up in a lot of master skills, e.g. a damage bonus only when wearing a two-handed sword.