Improvements for docker compose files
Solved a chicken-egg problem when setting up the docker-compose.prod.yml and certbot.
The problem was, that nginx did not initially start when we combined port 80 and 443 until the certificates were created.
So, when running the certbot the first time, we need to be able to access the port 80 - and we solve this by separating the server configs.