Implemented hot reload of configuration
When a config is changed and saved by an entity framework context:
* it will forward these changes to a IConfigurationChangeListener which is specific to the Persistence.EntityFramework.
* The implementation of that will
  * update the cached configuration objects and their parent objects, in case it's an collection
  * notify an instance of a ConfigurationChangeHandler which also existed before
  * The change handler will notify the change mediator (game logic), on which several logic objects will subscribe for changes

To make this work, especially updating the cached objects, I had to implement something like ICloneable<T> and Assignable<T>. I did this by implementing a code generator and a CloneableAttribute. The referencing of it might need some adaptions as it's probably only working in debug build right now.