Fixed invisible monsters
They occur when a freshly spawned monster starts a walk before it has been added to be observed by the player object.
In this case, ObserverToWorldViewAdapter.LocateableRemovedAsync leaves early, because the NewBucket is not null and the player is observing this bucket already.
Therefore, we can only start to walk, when the monster was safely added to the observing players before.