Refactored stats update
The IUpdateStatsPlugIn (view-plugin) implementations are responsible to inform the client about changes in certain stats like Health, Mana etc. Previously, only a fixed subset of the available stats were available for updates. Now we support to update all stats - the implementation can decide which stat it wants to send to the client.
Because some messages include more than one stat, and they can change in short amount of time, I implemented a mechanism which sends the update after a short delay of 16 ms. This results in less network messages.
Additionally, I removed the periodically sending of the messages, when nothing changed.