Fixed magic effect power up updates cause stat reset
Some skills update the max stat, e.g. Swell Life skill. When updating an existing effect, the max stat was reset to the normal value, and directly boosted again. When then the max value is above the current, the current gets capped to the lower max.
We prevent that now when the powerups are identical.