Fixed server info handler when running inside docker
The connect server has a mechanism to improve connectability when a client connects through an address which was not registered before.
When connecting to a all-in-one server which is running inside docker on windows, the local endpoint ip is set to the container ip within the docker network.
This IP is not available from outside. So in this case, we cannot return that back to the caller, because it will not be able to connect.
So to be sure, we don't use this "smart" logic when the server is running in docker anymore.